requirements.txt
influxdb==5.2.1
Flask==1.0.2
talisker==0.11.1
canonicalwebteam.http==1.0.1
